Add function to generate computer's choice
Added getComputerChoice() which is a function that generates the computers choice in rock paper scissors. It uses Math.Random() and Math.Floor() to generate a 0, 1, or 2. It then uses if and else if conditionals to map the number to either rock, paper, or scissors. Finally it returns this result as the computers choice.